# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "BenSerhat API",
    "version": "1.0.0",
    "description": "Kişisel site endpointleri"
  },
  "paths": {
...
    "/api/like": {
      "post": {
        "summary": "Beğeni ekle",
        "description": "",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slug": { "type": "string", "description": "Beğenilecek öğenin ID'si" },
                  "type": { "type": "string", "description": "Beğenilecek öğe türü (portfolio ,blog, comment, reply)" }
                },
                "required": ["blogId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Öğenin beğeni bilgilerini al",
        "description": "Belirtilen blogun, yorumun, yanıtın veya projenin beğeni bilgisini döner.",
        "parameters": [
          {
            "name": "slug",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Beğeni sayısı alınacak öğenin ID'si"
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Beğeni sayısı alınacak öğe türü (portfolio ,blog, comment, reply)"
          }
        ],
        "responses": {
          "200": {
            "description": "Mevcut beğeni sayısı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "likes": { "type": "number", "description": "Mevcut beğeni sayısı" },
                    "liked": { "type": "boolean", "description": "Kullanıcının beğenip beğenmediği", "default": false }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Beğeni kaldır",
        "description": "Belirtilen blog, yorum, yanıt veya projenin beğenisini kaldırır.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slug": { "type": "string", "description": "Beğenisi kaldırılacak öğenin ID'si" },
                  "type": { "type": "string", "description": "Beğenisi kaldırılacak öğe türü (portfolio ,blog, comment, reply)" }
                },
                "required": ["blogId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  }
}